# JITted RCaron