--- a/RCaron.Cli/RCaron.Cli.csproj
+++ b/RCaron.Cli/RCaron.Cli.csproj
@@ -9,12 +9,12 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="System.CommandLine" Version="2.0.0-beta3.22114.1" />
+        <PackageReference Include="System.CommandLine" Version="2.0.0-beta3.22114.1"/>
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\RCaron.FunLibrary\RCaron.FunLibrary.csproj" />
-      <ProjectReference Include="..\RCaron\RCaron.csproj" />
+        <ProjectReference Include="..\RCaron.FunLibrary\RCaron.FunLibrary.csproj"/>
+        <ProjectReference Include="..\RCaron\RCaron.csproj"/>
     </ItemGroup>
 
 </Project>