--- a/RCaron.Jit.Tests/ModuleInitializer.cs
+++ b/RCaron.Jit.Tests/ModuleInitializer.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.CompilerServices;
+
+namespace RCaron.Tests;
+
+public static class ModuleInitializer
+{
+    [ModuleInitializer]
+    public static void Initialize()
+    {
+        // throw new();
+        RCaron.Jit.Hook.EmptyMethod();
+    }
+}