--- a/RCaron.Jit.Tests/RCaron.Jit.Tests.csproj
+++ b/RCaron.Jit.Tests/RCaron.Jit.Tests.csproj
@@ -1,0 +1,35 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+
+        <IsPackable>false</IsPackable>
+        <EnablePreviewFeatures>True</EnablePreviewFeatures>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
+        <PackageReference Include="xunit" Version="2.4.2" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="coverlet.collector" Version="3.1.2">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+    </ItemGroup>
+
+    <ItemGroup>
+        <Compile Include="..\RCaron.Tests\*.cs" />
+        <ProjectReference Include="..\RCaron.Jit\RCaron.Jit.csproj">
+            <Properties>DefineConstants=RCARONJIT</Properties>
+        </ProjectReference>
+        <ProjectReference Include="..\RCaron\RCaron.csproj">
+            <Properties>DefineConstants=RCARONJIT</Properties>
+        </ProjectReference>
+    </ItemGroup>
+
+</Project>