After 2+ months of work it is done.
RCaron compiled to an expression tree.
Added benchmarks for it and the results are:
```rcaron
// firstly the code though
// SimpleMathOp: $h = 2 * (2 + 3);
// Fibbonaci:
$a = 0; $b = 1; $c = 0;

for($i = 0; $i < 2; $i++) {
    $c = $a + $b;
    $a = $b;
    $b = $c;
}
```
|                 Method |         Mean |        Error |       StdDev |       Median |  Ratio | RatioSD |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
|----------------------- |-------------:|-------------:|-------------:|-------------:|-------:|--------:|-------:|-------:|----------:|------------:|
|       SimpleMathOpFull |   3,386.7 ns |     15.58 ns |     22.84 ns |   3,392.8 ns |   1.00 |    0.00 | 0.4807 |      - |    3032 B |        1.00 |
|     SimpleMathOpParsed |     198.1 ns |      0.98 ns |      1.46 ns |     197.9 ns |   0.06 |    0.00 | 0.0153 |      - |      96 B |        0.03 |
|          FibbonaciFull |  12,564.2 ns |     66.47 ns |     99.49 ns |  12,558.4 ns |   3.71 |    0.04 | 1.4648 | 0.0153 |    9224 B |        3.04 |
|        FibbonaciParsed |   2,771.6 ns |     11.86 ns |     17.75 ns |   2,770.7 ns |   0.82 |    0.01 | 0.1984 |      - |    1248 B |        0.41 |
|   SimpleMathOpFull_Jit | 471,410.6 ns | 43,865.00 ns | 65,655.06 ns | 429,150.0 ns | 138.40 |   19.59 | 1.9531 | 1.4648 |   12978 B |        4.28 |
| SimpleMathOpParsed_Jit |     632.7 ns |     27.35 ns |     40.94 ns |     628.3 ns |   0.19 |    0.01 | 0.0076 |      - |      48 B |        0.02 |
|      FibbonaciFull_Jit | 863,414.6 ns | 57,770.53 ns | 84,679.29 ns | 854,418.2 ns | 254.85 |   23.91 | 3.9063 | 1.9531 |   25019 B |        8.25 |
|    FibbonaciParsed_Jit |     185.4 ns |      1.16 ns |      1.63 ns |     185.0 ns |   0.05 |    0.00 | 0.0420 |      - |     264 B |        0.09 |

Note the Fibbonaci test speeding up almost 15x from 2 771.6ns to 185.4ns and the allocations going down from 1248B to 264B(which is still this much because every variable is boxed to `object`).
Also, awkward. The SimpleMathOp test slowed down ~3.2x from 198.1ns to 632.7ns but the allocations getting split in half(96B to 48B). It could be because the is two dynamic operations there but when I inspect the Fibbonaci benchmark there is three. This does not make sense.
Also note the compiling for the JIT tests taking long. Yeah this is why we are keeping RCaron with an interpreted runtimes. I know you can just interpret an expression tree but having a dedicated interpreted sounds ~~fun~~ **painful**. Yeah may actually give up on that after thinking about it for a second. Would have to keep maintaining binders for both and awudwuiojduuiweiud. Will try interepreting expression tree.